Add TableName support to models
This commit adds a new "tablename" metadata tag that allows you to define a
custom table name for a given type.  Setting the metadata tag causes gorma to
generate a TableName function that returns the value of the metadata tag.  For
example, the following tag:

    Metadata("github.com/bketelsen/gorma#tablename", "admin.users")

causes gorma to generate the following function:

    func (m Users) TableName() string {
        return "admin.users"
    }