Refactor model struct generation
This commit makes `MakeModelDef` mimic `codegen.godef` as our entry point. It
calls `codegen.GoTypeDef` for subsequent attibute definitions.

The "duplicate ID removal hack" has been replaced by the `setupIDAttibution`
function.  It's job is to insert an ID attribute into the model or update an
existing one if it's already defined.

I also added a `genfuncs` map that we spin through to add extra fields and
comment the blocks as we go.