dsl: improve resource error message
When a user references a missing resource in a `Payload` DSL, include the
requested resource in the error message.  If they use an explicit type instead
of a string as the argument, the compiler will complain, so the error message in
question should only be printed when the user passes a string.